+++ tests/testthat/test-available_r.R
@@ -3,13 +3,13 @@ testthat::test_that("available_r lists all available r versions", {
     available_r(),
     c(
       "bleeding-edge", "frozen-edge", "r-devel", "bioc-devel",
-      "r-devel-bioc-devel", "latest-upstream",
+      "r-devel-bioc-devel", "latest-upstream", 
       "3.5.3", "3.6.0", "3.6.1", "3.6.2", "3.6.3",
       "4.0.0", "4.0.1", "4.0.2", "4.0.3", "4.0.4", "4.0.5",
       "4.1.0", "4.1.1", "4.1.2", "4.1.3",
       "4.2.0", "4.2.1", "4.2.2", "4.2.3",
       "4.3.0", "4.3.1", "4.3.2", "4.3.3",
-      "4.4.0", "4.4.1", "4.4.2"
+      "4.4.0", "4.4.1", "4.4.2", "4.4.3"
     )
   )
 })