+++ tests/testthat/test-rix_init.R
@@ -72,8 +72,6 @@ testthat::test_that("Snapshot test of rix_init(), create_missing, empty file", {
     rprofile_file
   }
 
-  testthat::announce_snapshot_file("find_rev/golden_Rprofile.txt")
-
   rprofile_con <- file(rprofile_file, open = "wb", encoding = "native.enc")
   on.exit(close(rprofile_con), add = TRUE)
 
@@ -105,7 +103,8 @@ testthat::test_that("Snapshot test of rix_init(), append", {
   rprofile_file <- paste0(path_env_nix, "/.Rprofile")
   rprofile_con <- file(rprofile_file, open = "a+", encoding = "native.enc")
 
-  writeLines(enc2utf8("This is in the original Rprofile"),
+  writeLines(
+    enc2utf8("This is in the original Rprofile"),
     rprofile_con,
     useBytes = TRUE
   )